refactor: :recycle: Start of refactoring and improvement of reuse, order and good practices
Start use of good practices for sorts files, also refactoring code for reuse code and optimization
