+++ package.json
@@ -13,6 +13,7 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.5.0",
     "react-scripts": "5.0.1",
+    "sass": "^1.58.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
