+++ src/App.js
@@ -1,20 +1,13 @@
+import "bootstrap/dist/css/bootstrap.min.css";
+import "bootstrap-icons/font/bootstrap-icons.css";
 import "./App.css";
-import Home from "./pages/home/Home.js";
-import TimersDetails from "./pages/timersDetails/TimersDetails.js";
-import NotFound from "./pages/notFound/NotFound";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
+import RoutesApp from "./routes/RoutesApp";
+
 
 function App() {
   return (
     <div className="App">
-      <BrowserRouter>
-        <Routes>
-          <Route path="/" element={<Home />} />
-          <Route path="/home" element={<Home />} />
-          <Route path="/timersDetails/:timerId" element={<TimersDetails />} />
-          <Route path="/*" element={<NotFound />} />
-        </Routes>
-      </BrowserRouter>
+      <RoutesApp />
     </div>
   );
 }
