+++ src/components/resetButton/ResetButton.js
@@ -5,7 +5,7 @@ const ResetButton = () => {
     <div className="resetButton-container">
       <div className="flex-item-button"></div>
       <Button variant="link" className="button resetButton">
-        <i class="bi bi-arrow-counterclockwise"></i>
+        <i className="bi bi-arrow-counterclockwise"></i>
       </Button>
     </div>
   );
