+++ src/components/timersTabs/TimersTabs.css
@@ -22,3 +22,49 @@
   0px 0px 3px 3px #fff,
   13px 8px 22px 1px rgb(5, 54, 36, 0.1);
 }
+
+.timersTabsBar {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  overflow: auto;
+}
+
+.timerLink{
+  text-decoration: none;
+  color: #02b875;
+}
+
+@media only screen and (min-width: 30px) 
+    and (max-width: 641px){
+      .timersTabsBar {
+        padding: 13% 0%;
+      }
+
+      .timerLinkButton {
+        margin: 0% 3%;
+      }
+}
+
+@media only screen and (min-width: 641px) 
+    and (max-width: 961px){
+      .timersTabsBar {
+        padding: 3% 0%;
+      }
+
+      .timerLinkButton {
+        margin: 0% 1%;
+      }
+}
+
+@media only screen and (min-width: 961px) 
+    /* and (max-width: 1081px) */
+{
+      .timersTabsBar {
+        padding: 2% 0%;
+      }
+
+      .timerLinkButton {
+        margin: 0% 1%;
+      }
+}
