+++ src/components/topBarTimersDetails/TopBarTimersDetails.css
@@ -3,6 +3,7 @@
   flex-direction: row;
   justify-content: space-evenly;
   align-items: center;
+  padding: 0% 3%;
 }
 
 .labelButton {
