+++ src/views/timersDetails/TimersDetails.css
@@ -5,12 +5,12 @@
   justify-content: space-evenly;
   align-self: center;
   color: #02b875;
-  /* border: black solid 1px; */
+  /* border: #02b875 solid 1px; */
 }
 
 .flex-container {
   align-self: center;
-  width: 70vmin;
+  width: 60vmin;
   display: flex;
   position: relative;
 }
@@ -72,4 +72,11 @@
 
 .mainCirclesContainer {
   position: relative;
+}
+
+.mainTimerContainer {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
 }
